% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{create_rnd_run_folder}
\alias{create_rnd_run_folder}
\title{Creates a random run folder and copies/creates model file and data stream file}
\usage{
create_rnd_run_folder(
  mod_file,
  base_run_folder,
  data_file = NULL,
  data_frame = NULL,
  nchar_folder = 10,
  sep = " ",
  na_char = ".",
  header = TRUE
)
}
\arguments{
\item{mod_file}{A valid path to an existing NONMEM model file.}

\item{base_run_folder}{A valid and existing base folder for NONMEM runs. In this folder a
random run folder will be created with a folder name of \code{nchar_folder} characters.}

\item{data_file}{A path or name of a NONMEM data stream file or \code{NULL}. See details for more
information.}

\item{data_frame}{A NONMEM data stream data.frame or \code{NULL}. See details for more
information.}

\item{nchar_folder}{The number of characters for the randomly named folder (must be 3-20).}

\item{sep}{If a \code{data_frame} should de serialized to a NONMEME data stream file: Column separator
of the file (defailt: whitepsace).}

\item{na_char}{If a \code{data_frame} should de serialized to a NONMEME data stream file: Character
for missing values (default: \code{.}).}

\item{header}{If a \code{data_frame} should de serialized to a NONMEME data stream file: \code{TRUE} if
column names should be present in the data stream file, else \code{FALSE}.}
}
\value{
A list with the following items:
\itemize{
\item \code{path} The full path to the run folder
\item \code{sdtab_file} The full path to the \code{sdtab} file (or \code{NA} if it could not be paraed from the
mod file).
\item \code{data_file} The full path to the data stream file.
\item \code{mod_file}  The full path to the NONMEM model file.
}
}
\description{
This function can be used to create new directories and run new models or existing models
with new data streams. In contrast to \code{\link[=create_run_folder]{create_run_folder()}}, tnis function
creates a randomly named folder inside \code{base_run_folder} to execute models. This function can
be useful in APIs or apps that need temporary run folders.
}
\details{
Data streams can either be provided as files (\code{data_file}) or
\code{data.frames} (\code{data_frame}).

The data stream file will be copied or serialized via the following rules:
\itemize{
\item The file name of the data stream is fetched via \code{\link[=mod_info]{mod_info()}}.
\item If \code{\link[=mod_info]{mod_info()}} could not parse the file, the base file name of \code{data_file} will be used.
\item If \code{data_file} and not \code{data.frame} is provided, \code{data_file} will to copied to the \code{run_folder}.
\item If \code{data_frame} is provided, the data.frame is serialized to the destination.
}
}
\examples{
\dontrun{
rf_info <- create_rnd_run_folder("mymod.mod", "/users/nmusers/myrunfolder/", data_frame = data)
rf_info <- create_rnd_run_folder("mymod.mod", "/users/nmusers/myrunfolder/", data_file = "test.csv")
}

}
\seealso{
\code{\link[=create_run_folder]{create_run_folder()}}, \code{\link[=remove_run_folder]{remove_run_folder()}}
}
