% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{create_run_folder}
\alias{create_run_folder}
\title{Creates a run folder and copies/creates model file and data stream file}
\usage{
create_run_folder(
  mod_file,
  run_folder,
  data_file = NULL,
  data_frame = NULL,
  create_folder = TRUE,
  sep = " ",
  na_char = ".",
  header = TRUE
)
}
\arguments{
\item{mod_file}{A valid path to an existing NONMEM model file.}

\item{run_folder}{A valid folder that exists or should be created (recursively) that
should act as a NONMEM run folder.}

\item{data_file}{A path or name of a NONMEM data stream file or \code{NULL}. See details for more
information.}

\item{data_frame}{A NONMEM data stream data.frame or \code{NULL}. See details for more
information.}

\item{create_folder}{If \code{run_folder} should be created (recursively). If \code{FALSE}, the function
assumes that this folder already exists.}

\item{sep}{If a \code{data_frame} should de serialized to a NONMEME data stream file: Column separator
of the file (defailt: whitepsace).}

\item{na_char}{If a \code{data_frame} should de serialized to a NONMEME data stream file: Character
for missing values (default: \code{.}).}

\item{header}{If a \code{data_frame} should de serialized to a NONMEME data stream file: \code{TRUE} if
column names should be present in the data stream file, else \code{FALSE}.}
}
\value{
A list with the following items:
\itemize{
\item \code{path} The full path to the run folder
\item \code{sdtab_file} The full path to the \code{sdtab} file (or \code{NA} if it could not be paraed from the
mod file).
\item \code{data_file} The full path to the data stream file.
\item \code{mod_file}  The full path to the NONMEM model file.
}
}
\description{
This function can be used to create new directories and run new models or existing models
with new data streams. The \code{run_folder} is created and model file and data stream files or
data_frames will be copied/created in this folder.
}
\details{
Data streams can either be provided as files (\code{data_file}) or
\code{data.frames} (\code{data_frame}).

The data stream file will be copied or serialized via the following rules:
\itemize{
\item The file name of the data stream is fetched via \code{\link[=mod_info]{mod_info()}}.
\item If \code{\link[=mod_info]{mod_info()}} could not parse the file, the base file name of \code{data_file} will be used.
\item If \code{data_file} and not \code{data.frame} is provided, \code{data_file} will to copied to the \code{run_folder}.
\item If \code{data_frame} is provided, the data.frame is serialized to the destination.
}
}
\examples{
\dontrun{
rf_info <- create_run_folder("mymod.mod", "/users/nmusers/myrunfolder/", data_frame = data)
rf_info <- create_run_folder("mymod.mod", "/users/nmusers/myrunfolder/", data_file = "test.csv")
}

}
\seealso{
\code{\link[=create_rnd_run_folder]{create_rnd_run_folder()}}, \code{\link[=remove_run_folder]{remove_run_folder()}}
}
