% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{mod_info}
\alias{mod_info}
\title{Reads information from a NONMEM model file}
\usage{
mod_info(mod_file)
}
\arguments{
\item{mod_file}{Valid path to a NONMEM model file. The file must be readable.

The function searches for \verb{$DATA} and \verb{$TABLE} entries in the model file.}
}
\value{
A list with the following fields:
\itemize{
\item \code{mod_file} the name of the NONMEM model file (full path)
\item \code{data_file} the name of the data stream input file or \code{NA} if not found.
\item \code{sdtab_file} the name of the output \code{sdtab} file or \code{NA} if not found.
}
}
\description{
Reads information from a NONMEM model file
}
\examples{
\dontrun{
infos <- mod_info("path/my_nm_model.mod")
}
}
