% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{nm_run}
\alias{nm_run}
\title{Runs a NONMEM session via system call}
\usage{
nm_run(
  mod_file,
  nm_starter,
  out_file_name = "nm_output",
  timeout = 0,
  debug = FALSE,
  check_data_file = TRUE
)
}
\arguments{
\item{mod_file}{A valid path to the NONMEM model file.}

\item{nm_starter}{The script of the NONMEM executable. See details for more infos.}

\item{out_file_name}{A name for the NONMEM output file.}

\item{timeout}{Timeout in seconds. If 0, no timeout is set.}

\item{debug}{If \code{TRUE}, \code{stdout} output will be displayed. The output will be captured and
print via \code{cat}.}

\item{check_data_file}{If \code{TRUE} the data stream file name will be parsed from the model file
and it is checked if this files exists in the folder of the model.}
}
\value{
A list with the following items:
\itemize{
\item \code{success} \code{TRUE} or \code{FALSE} that indicate if the execution was successful (this does not
indicate that the run was successful). If the run is timed out, \code{success} is \code{FALSE}.
\item \code{timeout} \code{TRUE} or \code{FALSE} if the run was timed out (only possible if \code{timeout > 0} was set).
\item \code{exec_time} The execution time in seconds.
\item \code{console_output} The console output of NONMEM as a single string.
\item \code{mod_file} The full path to the NONMEM model file that was executed.
\item \code{sdtab_file} The full path to the NONMEM output \code{sdtab} file (or \code{NA} if it could not be parsed
from the \code{mod} file).
\item \code{data_file} The full path to the NONMEM data stream file (or \code{NA} if it could not be parsed
from the \code{mod} file).
}
}
\description{
The NONMEM run will be executed within the folder of the model file.
The \code{nm_starter} is either a \code{bat} file on windows (e.g. \code{nm74/run/nmfe74.bat}) or a shell
script on linux (e.g. \code{nm751/run/nmfe75}).
}
\examples{
\dontrun{
run_ok <- nm_run("path/my_nm_model.mod", "/opt/nonmem/nm751/run/nmfe75")
}
}
