% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_sdtab}
\alias{read_sdtab}
\title{Reads a NONMEM sdtab file}
\usage{
read_sdtab(sdtab_file)
}
\arguments{
\item{sdtab_file}{A valid path to the \code{sdtab} file.}
}
\value{
A data.frame
}
\description{
Currently only one TABLE entry is supported!
The function tries to find if the table separator is \verb{,} or whitespace.
}
\examples{
\dontrun{
data <- read_sdtab("path/modelstdtab")
}
}
