% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{remove_run_folder}
\alias{remove_run_folder}
\title{Removes a folder, all files in the folder and all its subfolders}
\usage{
remove_run_folder(folder)
}
\arguments{
\item{folder}{Directory that should be removed}
}
\value{
The deleted paths (invisibly).
}
\description{
Removes a folder, all files in the folder and all its subfolders
}
\examples{
\dontrun{
remove_run_folder("myrunfolder")
}
}
\seealso{
\code{\link[=create_run_folder]{create_run_folder()}}, \code{\link[=create_rnd_run_folder]{create_rnd_run_folder()}}
}
