% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_nm}
\alias{test_nm}
\title{Test NONMEM Execution Capability}
\usage{
test_nm(nm_starter, debug = FALSE)
}
\arguments{
\item{nm_starter}{The script of the NONMEM executable. See details for more infos.}

\item{debug}{If \code{TRUE}, \code{stdout} output will be displayed. The output will be captured and
print via \code{cat}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{success}: \code{TRUE} if NONMEM run was successful, \code{FALSE} otherwise.
\item \code{msg}: Character string that contains NONMEM console output from the test run or error information.
}
}
\description{
This function tests the system's ability to execute NONMEM models by attempting to
start a NONMEM run via an internal test model. It checks
file paths, prepares a test environment, and captures any execution errors. The
function is useful for validating the setup of NONMEM in an R environment works.
}
\details{
The NONMEM run will be executed within the folder of the model file.
The \code{nm_starter} is either a \code{bat} file on windows (e.g. \code{nm74/run/nmfe74.bat}) or a shell
script on linux (e.g. \code{nm751/run/nmfe75}).
}
\examples{
\dontrun{
run_infos <- test_nm("/opt/nonmem/nm751/run/nmfe75")
}

}
